+++ src/Typin/Typin.Tests/ArgumentTests/UnsupportedArgumentConversionTests.cs
@@ -17,9 +17,9 @@
         }
 
         [Theory]
-        [InlineData("str-non-initializable-class")]
-        [InlineData("str-non-initializable-struct")]
-        [InlineData("str-enumerable-non-initializable")]
+        [InlineData("--str-non-initializable-class")]
+        [InlineData("--str-non-initializable-struct")]
+        [InlineData("--str-enumerable-non-initializable")]
         public async Task Property_of_custom_type_must_be_string_initializable_in_order_to_be_bound(string optionName)
         {
             // Arrange
@@ -27,14 +27,16 @@
                 .AddCommand<UnsupportedArgumentTypesCommand>();
 
             // Act
-            var (exitCode, _, stdErr) = await builder.BuildAndRunTestAsync(_output, new[]
+            var (exitCode, stdOut, stdErr) = await builder.BuildAndRunTestAsync(_output, new[]
             {
                 "cmd", optionName, "foobar"
             });
 
             // Assert
             exitCode.Should().NotBe(ExitCodes.Success);
+            stdOut.GetString().Should().BeNullOrWhiteSpace();
             stdErr.GetString().Should().NotBeNullOrWhiteSpace();
+            stdErr.GetString().Should().Contain("Can't convert");
         }
     }
 }